Fixes to improve IE7 Functionality
It wasn't HTML5 tags that broke the site - it was unsupported CSS, like
"inherit", and the lack of clear fix on containers. I've hard coded the
height using the html.lt-ie8 selector.

The dropdown menus still don't work because of z-index issues with IE7.
If we feel strongly, we can keep going, but the site will at be almost
functional like this in IE7  and won't look totally broken for  the
~0.8% using IE7.